Ajustes

+++ controller/deleteCategoriaController.php
@@ -10,15 +10,20 @@ $sql = "SELECT * FROM categorias WHERE id_categoria_pk = $idCategoria";
 $resultado1 = $conexion->query($sql);
 
 if ($resultado1->num_rows > 0) {
-  $sql2 = "DELETE FROM categorias WHERE id_categoria_pk = $idCategoria";
-  $resultado2 = $conexion->query($sql2);
+  $sqlV = "SELECT * FROM videojuegos WHERE id_categoria_fk = $idCategoria";
+  $resultado = $conexion->query($sqlV);
+  if($resultado->num_rows > 0){
+    echo 'error_3';
+  } else{  
+    $sql2 = "DELETE FROM categorias WHERE id_categoria_pk = $idCategoria";
+    $resultado2 = $conexion->query($sql2);
 
-  if (empty($resultado2)) {
-    echo 'error_2';
-  } else {
-    echo 'success';
+    if (empty($resultado2)) {
+      echo 'error_2';
+    } else {
+      echo 'success';
+    }
   }
-  
 } else {
   echo 'error_1';
 }
